requirements.txt
python-dotenv
cohere
panda
scikit-learn
numpy
altair
kagglehub